feat/ Regsiter and Search Kafka fucntioanlitites + tests for full workflow
